Rename tests to testUtils.
