+++ packages/core/tests/Reporter.test.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 import ansiEscapes from 'ansi-escapes';
-import { mockConsole, mockTool, mockRoutine } from '../src/tests';
+import { mockConsole, mockTool, mockRoutine } from '../src/testUtils';
 import Reporter, { testOnlyResetRestoreCursor } from '../src/Reporter';
 import ProgressOutput from '../src/outputs/ProgressOutput';
 import Output from '../src/Output';
